+++ common/common_entidad.cpp
@@ -1,9 +1,9 @@
 #include "common_entidad.h"
 
-Entidad::Entidad(uint8_t id, uint8_t pos_x, uint8_t pos_y): id(id), pos_x(pos_x), pos_y(pos_y) {}
+Entidad::Entidad(uint16_t id, uint16_t pos_x, uint16_t pos_y): id(id), pos_x(pos_x), pos_y(pos_y) {}
 
-uint8_t Entidad::get_id() { return id; }
+uint16_t Entidad::get_id() { return id; }
 
-uint8_t Entidad::get_pos_x() { return pos_x; }
+uint16_t Entidad::get_pos_x() { return pos_x; }
 
-uint8_t Entidad::get_pos_y() { return pos_y; }
+uint16_t Entidad::get_pos_y() { return pos_y; }
